dba_monitor/exporter/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
PyYAML
prometheus_client
